moved variable_names.rs from the top folder to the engine folder
