+++ pumpkin-solver/src/engine/variable_names.rs
